Introduce Framework-specific dependency
The WebConfigurationManager is a legacy library that is supported in .NET Framework, but which has no analog in .NET Core 3.x. As such, it will be supported on a project that has full access to the .NET Framework, but not on a project built entirely on .NET Core. This new `GetAppSettings()` method simply returns an array of any `<AppSettings />` that are part of the `web.config` file; if there aren't any, it will return an empty array.